+++ server.ts
@@ -1,9 +1,14 @@
 import 'source-map-support/register';
 import http from 'http';
-import chalk from 'chalk';
+import chalk, { Level } from 'chalk';
 
 import app from './server/app';
 import configs from "./server/helpers/configs";
+
+if(typeof configs.isTTY === "boolean") {
+  chalk.enabled = configs.isTTY;
+  chalk.level = configs.isTTY ? (chalk.level || Level.Basic) : Level.None;
+}
 
 let port = configs.port || 3000;
 if(process.env.DOCKERIZED) port = 80;